Implemented navigation to DrinkDetailsScreen from PairingResultsScreen
Implemented nested navigation to allow for stack navigation within the already existing tab navigation.
The user will be presented with the identified food item and suggested drink pairings.
This feature will allow for a user to click on a drink suggestion to get further information about any of the drinks.