Bug fix: Memory leak
Warning: Can't perform a React state update on an unmounted component. This is a no-op, but it indicates a memory leak in your application. To fix, cancel all subscriptions and asynchronous tasks in a useEffect cleanup function.

Solved by resetting form BEFORE completing action and possibly redirecting user to a different screen, which prevents an unmounted component from being updated.